Add unit conversions button to lab measurements
Consolidate measurement popup component
