--- app-frontend/src/app/components/map/labMap/labMap.component.js
@@ -1,15 +1,0 @@
-import labMapTpl from './labMap.html';
-const labMap = {
-    templateUrl: labMapTpl,
-    controller: 'LabMapController',
-    bindings: {
-        mapId: '@',
-        options: '<?',
-        initialCenter: '<?',
-        onClose: '&',
-        onCompareClick: '&',
-        comparing: '<'
-    }
-};
-
-export default labMap;
