--- app-frontend/src/app/components/map/measurementPopup/measurementPopup.component.js
@@ -1,14 +1,0 @@
-// Component code
-import measurementPopupTpl from './measurementPopup.html';
-
-const MeasurementPopupComponent = {
-    templateUrl: measurementPopupTpl,
-    bindings: {
-        delete: '&',
-        type: '<',
-        measurement: '<'
-    },
-    controller: 'MeasurementPopupController'
-};
-
-export default MeasurementPopupComponent;
