+++ app-frontend/src/assets/styles/sass/_shame.scss
@@ -2945,3 +2945,7 @@ rf-reclassify-table {
     pointer-events: all;
   }
 }
+
+.measurement-result {
+    margin-bottom: 1em;
+}
